+++ ipaas/src/components/IntegrationCreate/BallerinaCentralTokenPanel.tsx
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import { Accordion, AccordionDetails, AccordionSummary, Alert, Box, Button, Card, CircularProgress, InputAdornment, Link, Snackbar, Stack, TextField, Typography } from '@wso2/oxygen-ui';
+import { Accordion, AccordionDetails, AccordionSummary, Alert, Box, Button, Card, CircularProgress, InputAdornment, Link, Stack, TextField, Typography } from '@wso2/oxygen-ui';
 import { ArrowUpRight, ChevronDown, ChevronRight, Lock } from '@wso2/oxygen-ui-icons-react';
 import { useState, type JSX } from 'react';
 import { Link as NavigateLink } from 'react-router';
@@ -44,7 +44,7 @@ export default function BallerinaCentralTokenPanel({ org, tokenInput, onTokenInp
   const { data: tokenStatus, isLoading } = useBallerinaCentralToken();
   const saveToken = useSaveBallerinaCentralToken();
   const [instructionsOpen, setInstructionsOpen] = useState(false);
-  const [saveConfirmationOpen, setSaveConfirmationOpen] = useState(false);
+  const [alert, setAlert] = useState<{ type: 'success' | 'error'; message: string } | null>(null);
   const manageUrl = `/organizations/${org}/settings/package-registries`;
 
   const handleSave = () => {
@@ -52,106 +52,95 @@ export default function BallerinaCentralTokenPanel({ org, tokenInput, onTokenInp
     saveToken.mutate(tokenInput.trim(), {
       onSuccess: () => {
         onTokenInputChange('');
-        setSaveConfirmationOpen(true);
+        setAlert({ type: 'success', message: PANEL_COPY.saveSuccessMessage });
       },
+      onError: (err) => setAlert({ type: 'error', message: err.message }),
     });
   };
 
-  const saveSnackbar = (
-    <Snackbar open={saveConfirmationOpen} autoHideDuration={4000} onClose={() => setSaveConfirmationOpen(false)} anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}>
-      <Alert onClose={() => setSaveConfirmationOpen(false)} severity="success" variant="filled" sx={{ width: '100%' }}>
-        {PANEL_COPY.saveSuccessMessage}
-      </Alert>
-    </Snackbar>
-  );
-
   if (isLoading) return null;
-
-  if (tokenStatus?.configured) {
-    return saveSnackbar;
-  }
 
   return (
     <>
-      <Card variant="outlined" sx={{ p: 3, borderRadius: 2 }}>
-        <Stack direction="row" alignItems="center" gap={1.5} sx={{ mb: 0.5 }}>
-          <Typography variant="subtitle1" sx={{ fontWeight: 700 }}>
-            {PANEL_COPY.heading}
-          </Typography>
-          <Typography variant="body2" color="text.secondary">
-            {PANEL_COPY.optionalTag}
-          </Typography>
-        </Stack>
-        <Alert severity="warning" sx={{ mb: 2 }}>
-          <Stack direction="row" alignItems="center" flexWrap="wrap" gap={0.75}>
-            <Typography component="span" variant="body2">
-              {PANEL_COPY.warningPrefix}
-            </Typography>
-            <Link component={NavigateLink} to={manageUrl} underline="hover" sx={{ display: 'inline-flex', alignItems: 'center', gap: 0.5, fontWeight: 700 }}>
-              {PANEL_COPY.warningPathSegments.map((segment, index) => (
-                <Stack key={segment} direction="row" alignItems="center" gap={0.5} component="span">
-                  {index > 0 && <ChevronRight size={14} />}
-                  {segment}
-                </Stack>
-              ))}
-            </Link>
-          </Stack>
+      {alert && (
+        <Alert severity={alert.type} onClose={() => setAlert(null)} sx={{ mb: 2 }}>
+          {alert.message}
         </Alert>
-
-        <Accordion expanded={instructionsOpen} onChange={(_e, expanded) => setInstructionsOpen(expanded)} disableGutters sx={{ border: '1px solid', borderColor: 'divider', borderRadius: 1, mb: 2, '&:before': { display: 'none' } }}>
-          <AccordionSummary expandIcon={<ChevronDown size={16} />}>
-            <Typography variant="subtitle2" sx={{ fontWeight: 700 }}>
-              {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.heading}
+      )}
+      {!tokenStatus?.configured && (
+        <Card variant="outlined" sx={{ p: 3, borderRadius: 2 }}>
+          <Stack direction="row" alignItems="center" gap={1.5} sx={{ mb: 0.5 }}>
+            <Typography variant="subtitle1" sx={{ fontWeight: 700 }}>
+              {PANEL_COPY.heading}
             </Typography>
-          </AccordionSummary>
-          <AccordionDetails>
-            <Stack gap={3}>
-              <Box component="ol" sx={{ m: 0, pl: 2.5 }}>
-                {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.steps.map((step) => (
-                  <Typography key={step} component="li" variant="body2" sx={{ mb: 1 }}>
-                    {step}
-                  </Typography>
+            <Typography variant="body2" color="text.secondary">
+              {PANEL_COPY.optionalTag}
+            </Typography>
+          </Stack>
+          <Alert severity="warning" sx={{ mb: 2 }}>
+            <Stack direction="row" alignItems="center" flexWrap="wrap" gap={0.75}>
+              <Typography component="span" variant="body2">
+                {PANEL_COPY.warningPrefix}
+              </Typography>
+              <Link component={NavigateLink} to={manageUrl} underline="hover" sx={{ display: 'inline-flex', alignItems: 'center', gap: 0.5, fontWeight: 700 }}>
+                {PANEL_COPY.warningPathSegments.map((segment, index) => (
+                  <Stack key={segment} direction="row" alignItems="center" gap={0.5} component="span">
+                    {index > 0 && <ChevronRight size={14} />}
+                    {segment}
+                  </Stack>
                 ))}
-              </Box>
-              <Link href={BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkUrl} target="_blank" rel="noopener noreferrer" underline="hover" sx={{ display: 'inline-flex', alignItems: 'center', gap: 0.5, fontWeight: 600, width: 'fit-content' }}>
-                {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkLabel} <ArrowUpRight size={16} />
               </Link>
             </Stack>
-          </AccordionDetails>
-        </Accordion>
-
-        <Stack gap={1} sx={{ mb: 2 }}>
-          <Typography variant="subtitle2" sx={{ fontWeight: 700 }}>
-            {PANEL_COPY.accessTokenLabel}
-          </Typography>
-          <TextField
-            fullWidth
-            type="password"
-            value={tokenInput}
-            onChange={(e) => onTokenInputChange(e.target.value)}
-            placeholder={PANEL_COPY.tokenPlaceholder}
-            disabled={saveToken.isPending}
-            InputProps={{
-              startAdornment: (
-                <InputAdornment position="start">
-                  <Lock size={16} />
-                </InputAdornment>
-              ),
-            }}
-          />
-        </Stack>
-
-        {saveToken.isError && (
-          <Alert severity="error" sx={{ mb: 2 }}>
-            {saveToken.error.message}
           </Alert>
-        )}
 
-        <Button variant="contained" disabled={!tokenInput.trim() || saveToken.isPending} onClick={handleSave} startIcon={saveToken.isPending ? <CircularProgress size={14} /> : undefined} sx={{ alignSelf: 'flex-start' }}>
-          {PANEL_COPY.saveLabel}
-        </Button>
-      </Card>
-      {saveSnackbar}
+          <Accordion expanded={instructionsOpen} onChange={(_e, expanded) => setInstructionsOpen(expanded)} disableGutters sx={{ border: '1px solid', borderColor: 'divider', borderRadius: 1, mb: 2, '&:before': { display: 'none' } }}>
+            <AccordionSummary expandIcon={<ChevronDown size={16} />}>
+              <Typography variant="subtitle2" sx={{ fontWeight: 700 }}>
+                {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.heading}
+              </Typography>
+            </AccordionSummary>
+            <AccordionDetails>
+              <Stack gap={3}>
+                <Box component="ol" sx={{ m: 0, pl: 2.5 }}>
+                  {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.steps.map((step) => (
+                    <Typography key={step} component="li" variant="body2" sx={{ mb: 1 }}>
+                      {step}
+                    </Typography>
+                  ))}
+                </Box>
+                <Link href={BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkUrl} target="_blank" rel="noopener noreferrer" underline="hover" sx={{ display: 'inline-flex', alignItems: 'center', gap: 0.5, fontWeight: 600, width: 'fit-content' }}>
+                  {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkLabel} <ArrowUpRight size={16} />
+                </Link>
+              </Stack>
+            </AccordionDetails>
+          </Accordion>
+
+          <Stack gap={1} sx={{ mb: 2 }}>
+            <Typography variant="subtitle2" sx={{ fontWeight: 700 }}>
+              {PANEL_COPY.accessTokenLabel}
+            </Typography>
+            <TextField
+              fullWidth
+              type="password"
+              value={tokenInput}
+              onChange={(e) => onTokenInputChange(e.target.value)}
+              placeholder={PANEL_COPY.tokenPlaceholder}
+              disabled={saveToken.isPending}
+              InputProps={{
+                startAdornment: (
+                  <InputAdornment position="start">
+                    <Lock size={16} />
+                  </InputAdornment>
+                ),
+              }}
+            />
+          </Stack>
+
+          <Button variant="contained" disabled={!tokenInput.trim() || saveToken.isPending} onClick={handleSave} startIcon={saveToken.isPending ? <CircularProgress size={14} /> : undefined} sx={{ alignSelf: 'flex-start' }}>
+            {PANEL_COPY.saveLabel}
+          </Button>
+        </Card>
+      )}
     </>
   );
 }
