+++ ipaas/src/components/Settings/PackageRegistries/BallerinaCentralPanel.tsx
@@ -1,0 +1,138 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { Accordion, AccordionDetails, AccordionSummary, Alert, Box, Button, Card, CircularProgress, InputAdornment, Link, Skeleton, Stack, TextField, Typography } from '@wso2/oxygen-ui';
+import { ArrowLeft, ArrowUpRight, CheckCircle2, ChevronDown, Lock } from '@wso2/oxygen-ui-icons-react';
+import { useState, type JSX } from 'react';
+import { BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS } from '../../../constants/packageRegistries';
+import { useBallerinaCentralToken, useRemoveBallerinaCentralToken, useSaveBallerinaCentralToken } from '../../../hooks/useBallerinaCentralToken';
+import { formatDateTime } from '../../../utils/time';
+
+export default function BallerinaCentralPanel({ onBack }: { onBack: () => void }): JSX.Element {
+  const { data: tokenStatus, isLoading, isError, refetch } = useBallerinaCentralToken();
+  const saveToken = useSaveBallerinaCentralToken();
+  const removeToken = useRemoveBallerinaCentralToken();
+  const [tokenInput, setTokenInput] = useState('');
+  const [instructionsOpen, setInstructionsOpen] = useState(true);
+  const configured = !!tokenStatus?.configured;
+
+  const handleSave = () => {
+    if (!tokenInput.trim()) return;
+    saveToken.mutate(tokenInput.trim(), { onSuccess: () => setTokenInput('') });
+  };
+
+  return (
+    <Box>
+      <Button size="small" startIcon={<ArrowLeft size={14} />} onClick={onBack} sx={{ mb: 2 }}>
+        Back to Package Registries
+      </Button>
+      <Typography variant="h6" sx={{ mb: 0.5 }}>
+        Ballerina Central
+      </Typography>
+      <Typography variant="body2" color="text.secondary" sx={{ mb: 3 }}>
+        {configured ? 'This token is used by every build in the organization to pull private packages.' : 'Add a Ballerina Central access token so builds in this organization can pull private packages.'}
+      </Typography>
+
+      {isLoading ? (
+        <Skeleton variant="rounded" height={160} sx={{ maxWidth: 640 }} />
+      ) : isError ? (
+        <Alert severity="error" action={<Button onClick={() => refetch()}>Retry</Button>}>
+          Couldn&apos;t load the token status. Please try again.
+        </Alert>
+      ) : !configured ? (
+        <Stack gap={3} sx={{ maxWidth: 640 }}>
+          <Accordion expanded={instructionsOpen} onChange={(_e, expanded) => setInstructionsOpen(expanded)} disableGutters sx={{ border: '1px solid', borderColor: 'divider', borderRadius: 1, '&:before': { display: 'none' } }}>
+            <AccordionSummary expandIcon={<ChevronDown size={16} />}>
+              <Typography variant="subtitle1" sx={{ fontWeight: 700 }}>
+                {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.heading}
+              </Typography>
+            </AccordionSummary>
+            <AccordionDetails>
+              <Stack gap={3}>
+                <Box component="ol" sx={{ m: 0, pl: 2.5 }}>
+                  {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.steps.map((step) => (
+                    <Typography key={step} component="li" variant="body2" sx={{ mb: 1 }}>
+                      {step}
+                    </Typography>
+                  ))}
+                </Box>
+                <Link href={BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkUrl} target="_blank" rel="noopener noreferrer" underline="hover" sx={{ display: 'inline-flex', alignItems: 'center', gap: 0.5, fontWeight: 600, width: 'fit-content' }}>
+                  {BALLERINA_CENTRAL_TOKEN_INSTRUCTIONS.linkLabel} <ArrowUpRight size={16} />
+                </Link>
+              </Stack>
+            </AccordionDetails>
+          </Accordion>
+
+          <Stack gap={1}>
+            <Typography variant="subtitle2" sx={{ fontWeight: 700 }}>
+              Access token
+            </Typography>
+            <TextField
+              fullWidth
+              type="password"
+              value={tokenInput}
+              onChange={(e) => setTokenInput(e.target.value)}
+              placeholder="Paste your token here"
+              disabled={saveToken.isPending}
+              InputProps={{
+                startAdornment: (
+                  <InputAdornment position="start">
+                    <Lock size={16} />
+                  </InputAdornment>
+                ),
+              }}
+            />
+          </Stack>
+
+          {saveToken.isError && <Alert severity="error">{saveToken.error.message}</Alert>}
+
+          <Button variant="contained" disabled={!tokenInput.trim() || saveToken.isPending} onClick={handleSave} startIcon={saveToken.isPending ? <CircularProgress size={14} /> : undefined} sx={{ alignSelf: 'flex-start' }}>
+            Save token
+          </Button>
+        </Stack>
+      ) : (
+        <Stack gap={3} sx={{ maxWidth: 640 }}>
+          <Card variant="outlined" sx={{ p: 3, borderRadius: 2 }}>
+            <Stack direction="row" alignItems="center" gap={1} sx={{ color: 'success.main', mb: 3 }}>
+              <CheckCircle2 size={20} />
+              <Typography variant="subtitle1" sx={{ fontWeight: 700, color: 'success.main' }}>
+                Token configured
+              </Typography>
+            </Stack>
+            <Stack gap={0.5}>
+              <Typography variant="body2" color="text.secondary">
+                Added on
+              </Typography>
+              <Typography variant="body1" sx={{ fontWeight: 700 }}>
+                {formatDateTime(tokenStatus?.addedOn)}
+              </Typography>
+            </Stack>
+          </Card>
+
+          {removeToken.isError && <Alert severity="error">{removeToken.error.message}</Alert>}
+
+          <Stack direction="row" gap={1.5}>
+            <Button variant="outlined" color="error" disabled={removeToken.isPending} startIcon={removeToken.isPending ? <CircularProgress size={14} /> : undefined} onClick={() => removeToken.mutate()}>
+              Remove token
+            </Button>
+          </Stack>
+        </Stack>
+      )}
+    </Box>
+  );
+}
